Separere ut util for brev
