+++ apps/etterlatte-behandling/src/main/kotlin/brev/TilbakekrevingBrevService.kt
@@ -6,8 +6,6 @@ import kotlinx.coroutines.coroutineScope
 import no.nav.etterlatte.behandling.klienter.BrevApiKlient
 import no.nav.etterlatte.behandling.klienter.VedtakKlient
 import no.nav.etterlatte.brev.behandling.Soeker
-import no.nav.etterlatte.brev.behandling.erOver18
-import no.nav.etterlatte.brev.behandling.hentForelderVerge
 import no.nav.etterlatte.brev.behandling.mapAvdoede
 import no.nav.etterlatte.brev.behandling.mapInnsender
 import no.nav.etterlatte.brev.behandling.mapSoeker
@@ -19,18 +17,10 @@ import no.nav.etterlatte.brev.model.tilbakekreving.TilbakekrevingBrevInnholdData
 import no.nav.etterlatte.brev.model.tilbakekreving.TilbakekrevingDataNy
 import no.nav.etterlatte.brev.model.tilbakekreving.TilbakekrevingPeriodeDataNy
 import no.nav.etterlatte.grunnlag.GrunnlagService
-import no.nav.etterlatte.libs.common.behandling.BrevutfallDto
 import no.nav.etterlatte.libs.common.behandling.SakType
 import no.nav.etterlatte.libs.common.feilhaandtering.InternfeilException
 import no.nav.etterlatte.libs.common.feilhaandtering.UgyldigForespoerselException
-import no.nav.etterlatte.libs.common.grunnlag.Grunnlag
-import no.nav.etterlatte.libs.common.grunnlag.hentFoedselsnummer
-import no.nav.etterlatte.libs.common.grunnlag.hentSoekerPdlV1
 import no.nav.etterlatte.libs.common.objectMapper
-import no.nav.etterlatte.libs.common.person.UkjentVergemaal
-import no.nav.etterlatte.libs.common.person.Verge
-import no.nav.etterlatte.libs.common.person.Vergemaal
-import no.nav.etterlatte.libs.common.person.hentVerger
 import no.nav.etterlatte.libs.common.retryOgPakkUt
 import no.nav.etterlatte.libs.common.sak.SakId
 import no.nav.etterlatte.libs.common.tilbakekreving.JaNei
@@ -44,7 +34,6 @@ import no.nav.etterlatte.libs.common.vedtak.VedtakInnholdDto
 import no.nav.etterlatte.libs.ktor.route.logger
 import no.nav.etterlatte.libs.ktor.token.BrukerTokenInfo
 import no.nav.etterlatte.sak.SakService
-import no.nav.etterlatte.sikkerLogg
 import java.util.UUID
 import kotlin.math.absoluteValue
 
@@ -309,57 +298,3 @@ class TilbakeKrevingManglerVarsel :
         code = "TILBAKEKREVING_MANGLER_VURDERING_VARSEL",
         detail = "Kan ikke generere pdf uten at varsel er satt under vurdering",
     )
-
-fun hentVergeForSak(
-    sakType: SakType,
-    brevutfallDto: BrevutfallDto?,
-    grunnlag: Grunnlag,
-): Verge? {
-    val soekerPdl =
-        grunnlag.soeker.hentSoekerPdlV1()
-            ?: throw InternfeilException(
-                "Finner ikke søker i grunnlaget. Dette kan komme av flere ting, bl.a. endret ident på bruker. " +
-                    "Hvis dette ikke er tilfellet må feilen meldes i Porten.",
-            )
-
-    val verger =
-        hentVerger(
-            soekerPdl.verdi.vergemaalEllerFremtidsfullmakt ?: emptyList(),
-            grunnlag.soeker.hentFoedselsnummer()?.verdi,
-        )
-    return if (verger.size == 1) {
-        val vergeFnr = verger.first().vergeEllerFullmektig.motpartsPersonident
-        if (vergeFnr == null) {
-            logger.error(
-                "Vi genererer et brev til en person som har verge uten ident. Det er verdt å følge " +
-                    "opp saken ekstra, for å sikre at det ikke blir noe feil her (koble på fag). saken har " +
-                    "id=${grunnlag.metadata.sakId}. Denne loggmeldingen kan nok fjernes etter at løpet her" +
-                    " er kvalitetssikret.",
-            )
-            UkjentVergemaal()
-        } else {
-            // TODO: Hente navn direkte fra Grunnlag eller PDL
-            val vergenavn = "placeholder for vergenavn"
-
-            Vergemaal(
-                vergenavn,
-                vergeFnr,
-            )
-        }
-    } else if (verger.size > 1) {
-        logger.info(
-            "Fant flere verger for bruker med fnr ${grunnlag.soeker.hentFoedselsnummer()?.verdi} i " +
-                "mapping av verge til brev.",
-        )
-        sikkerLogg.info(
-            "Fant flere verger for bruker med fnr ${
-                grunnlag.soeker.hentFoedselsnummer()?.verdi?.value
-            } i mapping av verge til brev.",
-        )
-        UkjentVergemaal()
-    } else if (sakType == SakType.BARNEPENSJON && !grunnlag.erOver18(brevutfallDto?.aldersgruppe)) {
-        grunnlag.hentForelderVerge()
-    } else {
-        null
-    }
-}
